Added get student

+++ src/main/java/com/gritacademyapi/demo/repositories/StudentRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StudentRepository extends CrudRepository<Students, Integer> {
+public interface StudentRepository extends CrudRepository<Students, Long> {
 }
